feat(lin_reg): COMPLETED with gradient descent
Unfortunately, the gradient descent is not great to handle large
different between the intercept and the slope. As we know the learning
rate is used for both intercept and slope, and if those variable has a
big gap, it can be a stable learning rate for one of the variable but
can be unstable/bad to other that makes the calculation to divergence.

If i'm free, maybe i will update the calculation to handle this :memo:.